% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_through_genes_one_chrom.R
\name{cycle_through_genes_one_chrom}
\alias{cycle_through_genes_one_chrom}
\title{Process genes in one chrom}
\usage{
cycle_through_genes_one_chrom(
  assembled_dir_one_chrom = "tmpout/chrom_20/vcf_output"
)
}
\arguments{
\item{assembled_dir_one_chrom}{Dir path storing assembled files output by vcftools}
}
\value{
The exit status of future_walk function
}
\description{
Paralelly unite the three vcftool files of one gene into one rds for all genes on one chrom by calling the one gene processing function
unite_three_vcf_output_files_of_one_gene().
}
\examples{

# Process one chrom
tic("One tst chrom")
d <- "tmpout/chrom_20/vcf_output"
plan(multisession, workers = 12)
cycle_through_genes_one_chrom(
  assembled_dir_one_chrom = d
)
plan(sequential)
toc()

# Process multiple chroms
tic("Multiple chroms")
assembled_dir <- "tmpout"
plan(multisession, workers = 72)
dir_ls(assembled_dir, regexp = ".*chrom_[1-9].*") |>
  path("vcf_output") |>
    walk(function(chrom) {
      tic(chrom)
      cycle_through_genes_one_chrom(chrom)
      toc()
     })

plan(sequential)
toc()



}
