% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gene_bed.R
\name{extend_gene_bed}
\alias{extend_gene_bed}
\title{Extend the gene range for a bed file}
\usage{
extend_gene_bed(gene_bed_file, gene_extended_dir, extension)
}
\arguments{
\item{gene_bed_file}{The original gene bed file path}

\item{gene_extended_dir}{The directory to store the extended bed file}

\item{extension}{An integer, extend equal amount at both sides, start - extension and end + extension}
}
\value{
The path to the extended bed file
}
\description{
The bed file should have four columns, "chrom", "start", "end", "gene", but has no column names.
}
\examples{

# Store the extended gene bed file in a temp directory
gene_extended_dir <-
  path_temp("gene_extended") |>
  dir_create()
# Function to extent a gene bed file
gene_dir=fs::path_package("PLEIOVAR", "extdata", "gene")
gene_bed_file <- path(gene_dir, "chrom_20.bed")
extension <- 50000

extend_gene_bed(gene_bed_file, gene_extended_dir, extension)

}
