% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genewise_genotype.R
\name{get_genewise_genotype}
\alias{get_genewise_genotype}
\title{Get genewise genotype via vcftools}
\usage{
get_genewise_genotype(
  vcf_dir = fs::path_package("PLEIOVAR", "extdata", "vcf"),
  gene_dir = fs::path_package("PLEIOVAR", "extdata", "gene"),
  out_dir = "result",
  job_num = 30,
  extension = 50000
)
}
\arguments{
\item{vcf_dir}{path to the dir including zipped chrom-wise vcf files with "chr[om_]*\\d+" in their names}

\item{gene_dir}{path to the dir including bed files with naming format chrom_${i}.bed.  One bed file each chrom.}

\item{out_dir}{path to the dir to store the outputs, a subdir, chrom_${i}/vcf_output, will created for each chrom.}

\item{job_num}{Jobs to run parallely, passed down to GNU parallel utility}

\item{extension}{Integer, the nucleotide amount to be extended at both ends}
}
\value{
The status of pwalk
}
\description{
It sequentially loops over the chroms under vcf_dir, pulls the variants for each gene in the bed file
of the corresponding chrom located in gene_dir, which is parallelized using GNU parallel utility, and output to
out_dir/chrom_i/vcf_output. i is the chrom number.  One gene generates three files, gene.012, gene.012.indv, and gene.012.pos.
}
\examples{

# Get the dir path
library(fs)
vcf_dir <- path_package("PLEIOVAR", "extdata", "vcf")
gene_dir <- path_package("PLEIOVAR", "extdata", "gene")
out_dir <- "result"
job_num <- 30
extension <- 50000
# Pull variants from vcf files
get_genewise_genotype(vcf_dir, gene_dir, out_dir, job_num, extension)


}
