% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genewise_genotype_via_vcftools.R
\name{get_genewise_genotype_via_vcftools}
\alias{get_genewise_genotype_via_vcftools}
\title{Pull variants by gene from vcf files}
\usage{
get_genewise_genotype_via_vcftools(
  vcf_dir = fs::path_package("PLEIOVAR", "extdata", "vcf"),
  gene_dir = fs::path_package("PLEIOVAR", "extdata", "gene"),
  out_dir = "result"
)
}
\arguments{
\item{vcf_dir}{path to the dir including zipped chrom-wise vcf file with naming format Sardinia.b37.ss2120.FAref.impv4.chr${i}.vcf.gz}

\item{gene_dir}{path to the dir including bed files with naming format chrom_${i}.bed.  One bed file each chrom.}

\item{out_dir}{path to the dir to store the outputs, a subdir, chrom_${i}/vcf_output, will created for each chrom.}
}
\value{
The return value is an error code (0 for success). See the "value" section of system2 for details.
}
\description{
It wraps a shell script, which sequentially loops over the chroms under vcf_dir, pulls the variants for each gene present in the bed file
of the corresponding chrom under gene_dir -- this step is parallelized using GNU parallel utility, and save the output to the subdirectory,
chrom_i/vcf_output, under out_dir. i is the chrom number.  One gene generates three files, gene.012, gene.012.indv, and gene.012.pos.
}
\examples{
# Get the dir path
library(fs)
vcf_dir <- path_package("PLEIOVAR", "extdata", "vcf")
gene_dir <- path_package("PLEIOVAR", "extdata", "gene")
out_dir <- "result"
# Pull variants from vcf files
get_genewise_genotype_via_vcftools(vcf_dir, gene_dir, out_dir)
}
