% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_snp.R
\name{pc_snp}
\alias{pc_snp}
\title{Generate principle components for SNPs within genes}
\usage{
pc_snp(
  infolder = system.file("extdata", "pc_snp", package = "PLEIOVAR"),
  outfolder = "dev/func_output",
  OPF_snp = 0.75,
  freq_cut = 0.005,
  block_name = "blocktest"
)
}
\arguments{
\item{infolder}{This is the folder where the input files are located}

\item{outfolder}{This is the folder where the output results will be written to}

\item{OPF_snp}{variance explained parameter (0.75 in our case) for PC-SNP dimension reduction}

\item{freq_cut}{minor allele frequency cutoff for a SNP to be used for PC-SNP (in our case this is 0.005)}

\item{block_name}{the number of the block file}
}
\value{
No explicit returns specified yet
}
\description{
This programs aims at selecting the SNPS from the SardiNIA vcf file which correspond to a specific gene region
After preprocessing was performed fom previous steps, we have a gene region dosage file for each gene
block files (one for each CPU when using multiple jobs in parallel) contain a list of all the genes in the block in UCSC format (chr,start,end,gene)
So, in summary, these block files are really subsets of the file All_Genes_hg19 which contains all the genes
Here is an example for the block file number 27 (out of a total of 64 blocks)
7 137074384 137531609 DGKI
-7 137597556 137686847 CREB3L2
-7 137638093 137642712 LOC100130880
-7 137761177 137803050 AKR1D1
Of the block file described above, we use only the gene name so we can locate the corresponding gene region file {gene name}_assembled
Next, for each gene region (in the block) , read the "assembled" file for that gene and then we generate the PC-SNPs for each gene.
-Next, we perform a dimension reduction which results from a much smaller set of PC-SNPs than the number of SNPs in that gene region
}
\examples{

infolder <- system.file("extdata", "pc_snp", package = "PLEIOVAR")
outfolder <- "dev/func_output"
OPF_snp <-  0.75
freq_cut <-  0.005
block_name <-  "blocktest"
pc_snp(infolder, outfolder, OPF_snp, freq_cut, block_name)

}
