% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_snp_one_gene.R
\name{pc_snp_one_gene}
\alias{pc_snp_one_gene}
\title{Generate pc_snp for one gene}
\usage{
pc_snp_one_gene(
  gene = "gene_20_2",
  assembled_dir = "result/chrom_20/assembled_file",
  pc_snp_dir = "result/chrom_20/pc_snp",
  freq_cut = 0.005,
  OPF_snp = 0.75
)
}
\arguments{
\item{gene}{Gene name}

\item{assembled_dir}{The directory containing assembled rds files}

\item{pc_snp_dir}{The directory to which PCA analysis results will be written}

\item{freq_cut}{Minor allele frequency cutoff for a SNP to be used for PC-SNP (in our case this is 0.005)}

\item{OPF_snp}{Variance explained parameter (0.75 in our case) for PC-SNP dimension reduction}
}
\value{
The return value of the function writing PCA results out
}
\description{
Reads the assembled file from assembled dir, which contains the genotype information for each gene in the file
"gene_name.rds", run PCA, and write output to the three subdirectories, eigenvalue, loading, and pc_snp_score, under
pc_snp_dir.
}
\examples{

pc_snp_one_gene(
  gene = "gene_20_2",
  assembled_dir = "result/chrom_20/assembled_file",
  pc_snp_dir ="result/chrom_20/pc_snp",
  freq_cut =  0.005,
  OPF_snp = 0.75)


}
