% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_trait.R
\name{pc_trait}
\alias{pc_trait}
\title{Generate the principal components for phenotypic traits}
\usage{
pc_trait(
  pheno_file = "extdata/pheno/mainpheno.dat",
  pc_trait_dir = "result/pc_trait",
  OPF_trait = 0.99
)
}
\arguments{
\item{pheno_file}{phenotype data. The first column should be named "ID", followed by one phenotype per column}

\item{pc_trait_dir}{The outputs are saved to this directory}

\item{OPF_trait}{The threshold for filtering cumulative eigen values, i.e., the percentage of explained variances}
}
\value{
The return value of the function saving trait PC components
}
\description{
Read in phenotype data, conduct principal component analysis, save out the loadings and the scores.
}
\examples{

pheno_file <- fs::path_package("PLEIOVAR", "extdata", "pheno", "mainpheno.dat")
# pc_trait_dir <- fs::path_package("PLEIOVAR", "result", "pc_trait")
pc_trait_dir <- fs::path("~/project/PLEIOVAR/PLEIOVAR_rpkg", "result", "pc_trait")
OPF_trait <-  0.99
pc_trait(pheno_file = pheno_file, pc_trait_dir = pc_trait_dir, OPF_trait = OPF_trait)

}
