% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_three_vcf_output_files_of_one_gene.R
\name{unite_three_vcf_output_files_of_one_gene}
\alias{unite_three_vcf_output_files_of_one_gene}
\title{Unite the three output files of vcftools of one gene}
\usage{
unite_three_vcf_output_files_of_one_gene(
  assembled_dir_one_chrom = "../../assembled/variants_by_gene/chrom_21",
  gene = "ABCC13"
)
}
\arguments{
\item{assembled_dir_one_chrom}{Dir path storing assembled files output by vcftools}

\item{gene}{character, the name of a single gene

The assembled dir path should follow the rule parent_dirs/vcf_output. The three files of a gene will be located according to the
naming rule mentioned above. The united file is saved as rds at parent_dirs/rds and create this dir if it has not existed. The genes
with zero SNPs are not converted to rds.}
}
\value{
The status of write_rds function
}
\description{
Three files are output by vcftools for each gene, gene.012 containing individual x SNP matrix, gene.012.indv containing
individual IDs in one column, gene.012.pos containing chrom and SNP coordinate in two columns
}
\examples{
# The rds file is saved at "../../assembled/variants_by_gene/chrom_tst/rds"
d <- path_package("PLEIOVAR") |>
  path_dir() |>
  path("../assembled/variants_by_gene/chrom_tst/vcf_output")

unite_three_vcf_output_files_of_one_gene(
    assembled_dir_one_chrom = d,
    gene = "ABCC13"
  )

# Gene "MIR155" has no SNPs, so no rds file saved
unite_three_vcf_output_files_of_one_gene(
    assembled_dir_one_chrom = d,
    gene = "MIR155"
  )


}
