% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_three_vcf_output_files_of_one_gene.R
\name{unite_three_vcf_output_files_of_one_gene}
\alias{unite_three_vcf_output_files_of_one_gene}
\title{Unite the three output files of vcftools for one gene}
\usage{
unite_three_vcf_output_files_of_one_gene(
  assembled_dir_one_chrom = "tmpout/chrom_20/vcf_output",
  gene = "gene_20_2"
)
}
\arguments{
\item{assembled_dir_one_chrom}{Dir path storing assembled files generated by vcftools}

\item{gene}{character, the name of a single gene

The assembled dir path should follow the format chrom_i/vcf_output. The three files of a gene will be located according to the
naming rule mentioned above. The united file is saved as rds at chrom_i/rds and create this dir if it has not existed. The genes
with zero SNPs are not converted to rds.}
}
\value{
The status of write_rds function
}
\description{
Three files are generated by vcftools for each gene, gene.012 containing individual x SNP matrix, gene.012.indv containing
individual IDs in one column, gene.012.pos containing chrom and SNP coordinate in two columns. This function combines the
three files and save it as a rds file under "rd" folder, which will be created if not pre-exists, at the same level of
assembled dir, i.e, if assembled dir is chrom_20/vcf_output, the united rds file goes to chrom_20/rds. Only genes containing
non-zero variants are kept.
}
\examples{
# The rds file is saved at "../../assembled/variants_by_gene/chrom_tst/rds"

d <- "tmpout/chrom_20/vcf_output"
unite_three_vcf_output_files_of_one_gene(
    assembled_dir_one_chrom = d,
    gene = "gene_20_2"
  )

# Gene "MIR155" has no SNPs, so no rds file saved
unite_three_vcf_output_files_of_one_gene(
    assembled_dir_one_chrom = d,
    gene = "MIR155"
  )


}
